update: remove utils & update jsdoc
